refactor: move dist_alerts `get` to new router
